consumer/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
kafka-python
